+++ jindex-rs/src/jni/jni_indexed_field.rs
@@ -1,5 +1,5 @@
 use jni::objects::JObject;
-use jni::sys::{jobject, jshort, jstring};
+use jni::sys::{jint, jobject, jstring};
 use jni::JNIEnv;
 
 use crate::class_index::{IndexedClass, IndexedField};
@@ -33,11 +33,11 @@ pub unsafe extern "system" fn Java_com_github_tth05_jindex_IndexedField_getName(
 pub unsafe extern "system" fn Java_com_github_tth05_jindex_IndexedField_getAccessFlags(
     env: JNIEnv,
     this: jobject,
-) -> jshort {
+) -> jint {
     let indexed_field =
         get_field_with_id::<IndexedField>(env, this, &cached_field_ids().indexed_field_pointer_id);
 
-    indexed_field.access_flags() as jshort
+    indexed_field.access_flags() as jint
 }
 
 #[no_mangle]
